+++ src/index.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
-import * as serviceWorker from './serviceWorker';
+import * as serviceWorker from './service-worker';
 import ReduxWrapper from 'wrapper/ReduxWrapper';
 import Router from 'wrapper/Router';
 import 'styles/tailwind.out.css';
