+++ src/utilities/index.ts
@@ -7,13 +7,13 @@ export {default as environment} from './constants/environment';
 export {default as formOptions} from './constants/form-options';
 export {default as templates} from './constants/templates';
 
-export {default as formatters} from './formatters';
-export {default as dispatchers} from './dispatchers';
-export {default as validators} from './validators';
-export {default as hints} from './hints';
+export {default as formatters} from './form-helpers/formatters';
+export {default as dispatchers} from './redux-helpers/dispatchers';
+export {default as validators} from './form-helpers/validators';
+export {default as hints} from './form-helpers/hints';
 
-export {default as fetchSurveys} from './surveyCommunication';
-export {default as generateValidOAuthToken} from './jwtEncryption';
+export {default as fetchSurveys} from './ajax-helpers/fetching';
+export {default as generateValidOAuthToken} from './auth-helpers/jwt-handling';
 
-export {authPostRequest} from './axiosClients';
-export {surveyGetRequest} from './axiosClients';
+export {authPostRequest} from './ajax-helpers/axios-clients';
+export {surveyGetRequest} from './ajax-helpers/axios-clients';
